First attempt at Rectangle. Fails ...
running -test
osmart@fuji01vm:/mnt/hgfs/vmshare/msc/pij/pij/day03$ java E13Rectangle  -test
E13Rectangle program reads in X and Y coordinates for two points and outputs
  the area and perimeter of a rectangle where both points are opposite corners.
-test procedure, use hard coded points:
Exception in thread "main" java.lang.NullPointerException
	at Rectangle.storePoints(E13Rectangle.java:64)
	at E13Rectangle.main(E13Rectangle.java:97)

Because although the default constructor initializes the myRectangle
object the Points upLeft and downRight are not initialized.
So when we try to store anything in the we get a "NullPointerException"
because there is no where to store anything.

Need a constructor?!?!