Start of the ListIntSet implementation. Test output ->
-> Test output for linked list test just a load of print
debug so far!

$ java TestIntSets
Run tests on TreeIntSet:
add 7 to the set
add 9 to the set
add 7 to the set again!
print the set contents: 7, 9
verbose test whether set contains 9:
	contains checks 9 against 7
	contains checks 9 against 9
	Result set does contain 9
extend set to: 7, 3, 1, 9, 8, 49, 13, 29, 22, 21, 19, 14, 18, 23, 27, 26, 45, 35, 34, 33, 37, 41, 39, 43, 44, 50
verbose test whether set contains 19:
	contains checks 19 against 7
	contains checks 19 against 9
	contains checks 19 against 49
	contains checks 19 against 13
	contains checks 19 against 29
	contains checks 19 against 22
	contains checks 19 against 21
	contains checks 19 against 19
	Result set does contain 19
verbose test whether set contains 28:
	contains checks 28 against 7
	contains checks 28 against 9
	contains checks 28 against 49
	contains checks 28 against 13
	contains checks 28 against 29
	contains checks 28 against 22
	contains checks 28 against 23
	contains checks 28 against 27
	Result set does not contain 28

Re-Run the same tests on ListIntSet:
add 7 to the set
debug ListIntSet.add() to be written
add 9 to the set
debug ListIntSet.add() to be written
add 7 to the set again!
debug ListIntSet.add() to be written
print the set contents: debug ListIntSet.toString() to be written
verbose test whether set contains 9:
debug ListIntSet.containsVerbose() to be written
	Result set does not contain 9
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
debug ListIntSet.add() to be written
extend set to: debug ListIntSet.toString() to be written
verbose test whether set contains 19:
debug ListIntSet.containsVerbose() to be written
	Result set does not contain 19
verbose test whether set contains 28:
debug ListIntSet.containsVerbose() to be written
	Result set does not contain 28